refactor: removed unused code, renamed package
